fix: renaming files and exports
