+++ apps/docs/content/docs/develop/(hooks)/use-theme.client.tsx
