+++ apps/docs/content/docs/develop/(hooks)/use-theme.mdx
@@ -1,5 +1,5 @@
 ---
-title: useThemeToggler
+title: useTheme
 description: A hook and components for managing and toggling themes dynamically.
 ---
 
@@ -9,7 +9,7 @@ import {
   ThemeProvider,
   useTheme,
   SampleThemeToggler,
-} from "./use-theme-toggler.client";
+} from "./use-theme.client";
 import { TypeTable } from "fumadocs-ui/components/type-table";
 import { Tab, Tabs } from "fumadocs-ui/components/tabs";
