+++ apps/docs/content/docs/develop/(hooks)/use-theme.ms.mdx
@@ -1,5 +1,5 @@
 ---
-title: useThemeToggler
+title: useTheme
 description: The Toggle component is a versatile switch that allows users to toggle between two states.
 ---
