+++ apps/docs/content/docs/develop/meta.en.json
@@ -52,7 +52,7 @@
     "(components)/tooltip",
 
     "---Hooks ---",
-    "(hooks)/use-theme-toggler"
+    "(hooks)/use-theme"
   ],
   "root": true
 }
