+++ packages/react/src/hooks/index.ts
@@ -2,4 +2,4 @@ export { useCallbackRef } from "./use-callback-ref";
 export { useControllableState } from "./use-controllable-state";
 export { default as useToast } from "./use-toast";
 export { usePagination } from "./use-pagination";
-export { useTheme } from "./use-theme-toggler";
+export { useTheme } from "./use-theme";
